warnings: fix the three first-party Teensy build warnings
The device build (-Wall -Wextra) emitted three first-party warnings, all in
code paths the native -Werror build never compiles (device NDEBUG / uint16
promotion), so they rode green CI:

- core/platform.h: HS_CHECK(cond) with no message expanded to
  check_fail(..., "") — a zero-length gnu_printf format (-Wformat-zero-length).
  Switch the macro to __VA_OPT__ plus a 3-arg check_fail overload that formats
  an empty message via ("%s", ""), so no empty format literal is ever passed;
  printf format-checking is preserved when a message IS given.
- hardware/pov_sync.h: `bound` is read only inside assert(), which is compiled
  out under the device's NDEBUG -> [[maybe_unused]].
- core/animation.h: active_count is uint16_t, so active_count - 1 promotes to
  int and trips -Wsign-compare against size_t i -> cast the result to size_t.

Clears the now-empty first-party warning baseline.

+++ core/animation.h
@@ -637,7 +637,7 @@ public:
       for (size_t i = 0; i < active_count; ++i) {
         bool dead = step_particle(pool[i], max_delta);
         if (dead) {
-          if (i != active_count - 1) {
+          if (i != static_cast<size_t>(active_count - 1)) {
             pool[i] = pool[active_count - 1];
           }
           active_count--;
